requirements.txt
asgiref==3.8.1
attrs==25.1.0
black==23.11.0
click==8.1.8
dj-database-url==2.1.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.26.5
et_xmlfile==2.0.0
factory-boy==3.3.0
Faker==37.0.0
flake8==6.1.0
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.9.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
whitenoise==6.6.0
